Start generalisation [skip actions]
