+++ autode/calculations/__init__.py
@@ -1,0 +1,3 @@
+from autode.calculations.calculation import Calculation
+
+__all__ = ["Calculation"]
