+++ autode/wrappers/ri.py
@@ -1,5 +1,5 @@
 from autode.wrappers.keywords import RI
 
 rijcosx = RI(name='rijcosx',
-             doi_list=['10.1016/j.chemphys.2008.10.036'],
+             DOIs=['10.1016/j.chemphys.2008.10.036'],
              orca='RIJCOSX')
